==> fisher/methylation/7_Gene_expression_analysis.ipynb <==
# <span style="color:#ff1414"> BEDtools analysis. </span>

This is a script to answer research questions outlined elsewhere. In summary, this script:

1. compares methylation results between different methylation-callers, and between different methylation sequencing methods.

2. compares methylation between genes and non-gene regions

3. compares methylation between transposons and non-repetitive regions

4. compares transposons and genes


Note:
- PB/pb = PacBio
- ONT/ont = Oxford Nanopore Technology
- NP = Nanopolish

import pybedtools
import scipy

import matplotlib.patches as mpatches
import pandas as pd
import numpy as np # need for  stats

from scipy.stats import wilcoxon
import matplotlib.pyplot as plt
from matplotlib_venn import venn2

# load modules
import os
import glob
import pprint
from pybedtools import BedTool
from scipy.stats import spearmanr

#First we need to define the base dirs
DIRS = {}
DIRS['BASE2'] = '/home/anjuni/analysis'
DIRS['FEATURES'] = os.path.join(DIRS['BASE2'], 'coverage', 'feature_files')
DIRS['RAND'] = os.path.join(DIRS['BASE2'], 'coverage', 'randomisation')
DIRS['COVERAGE'] = os.path.join(DIRS['BASE2'], 'coverage')
DIRS['RNA'] = os.path.join(DIRS['BASE2'], 'rna_counts')
DIRS['edgeR'] = os.path.join(DIRS['RNA'], 'edgeR_output')
DIRS['TRIALS'] = os.path.join(DIRS['RNA'], 'trials_tsv')
DIRS['MEAN_STD'] = os.path.join(DIRS['RNA'], 'average_and_stdev')
DIRS['FIGURES'] = os.path.join(DIRS['BASE2'], 'figures')

DIRS['GENE'] = os.path.join(DIRS['COVERAGE'], 'gene_level')
DIRS['GENE_BODY'] = os.path.join(DIRS['GENE'], 'gene_body')
DIRS['BOTH_U_D'] = os.path.join(DIRS['GENE'], 'both_upstream_downstream')
DIRS['DOWN_STR'] = os.path.join(DIRS['GENE'], 'downstream')
DIRS['UP_STR'] = os.path.join(DIRS['GENE'], 'upstream')
DIRS['TSS'] = os.path.join(DIRS['GENE'], 'tss_6mA_only')
DIRS['TE'] = os.path.join(DIRS['GENE'], 'te')
DIRS['COMPARE'] = os.path.join(DIRS['BASE2'], 'gene_te_comparison')
DIRS['GENE_TE'] = os.path.join(DIRS['COMPARE'], 'output', 'both')

#Quick chech if directories exist
for value in DIRS.values():
    if not os.path.exists(value):
        print('%s does not exist' % value)

## <span style='color:#15c66e'> 7. Analysing gene expression files. <span/>

### <span style='color:#15c66e'> 7.A Only genes. <span/>

#### <span style='color:#a347ff'> 7.A.1 Making dataframe with methylation + expression data.<span/>

# Prepare RNA-seq inputs
rnaseq_list = [os.path.join(DIRS['MEAN_STD'], 'Pst_104E_v13_h_gene_rpkm_average.tsv'), \
               os.path.join(DIRS['MEAN_STD'], 'Pst_104E_v13_p_gene_rpkm_average.tsv')]

rnaseq_list

# make heading list for converting the methylation coverage bed file to dataframe :)
headings = ['contig', 'start', 'stop', 'gene_ID', 'overlap_count', 'overlap_fraction']

# Convert list to dict of dataframes
def file_name_dict(file_list):
    """Outputs a dictionary of input file paths for a given list of input file paths."""
    file_dict = {}
    for file in file_list:
        file_dict[file.split('/')[-1]] = file
    return file_dict

def df_dict_from_fn_dict(file_dict, header_row, header_list):
    """Outputs a dictionary of dataframes from a dictionary of file names."""
    df_dict = {}
    for key, value in file_dict.items():
        df = pd.read_csv(value, sep='\t', header=header_row, names=header_list)
        df = df.loc[:,['gene_ID', 'overlap_fraction']].sort_values('gene_ID', ascending=True)
        df_dict[key] = df
    return df_dict

def df_dict_from_fn_list(file_list, header_row, header_list):
    """Outputs a dictionary of dataframes from a dictionary of file names."""
    df_dict = {}
    for file in file_list:
        df = pd.read_csv(file, sep='\t', header=header_row, names=header_list)
        df = df.loc[:,['gene_ID', 'overlap_fraction']]#.sort_values('gene_ID', ascending=True)  # sorting alphabetically messes it up, so I won't do it here
        df_dict[fn.split('/')[-1]] = df
        df_dict[file.split('/')[-1]] = df
    return df_dict

#Make the dataframes from part 6.
h_contig_df = pd.read_csv(os.path.join(DIRS['FIGURES'], 'coverage', 'coverage_df_h.tsv'), sep='\t', )
p_contig_df = pd.read_csv(os.path.join(DIRS['FIGURES'], 'coverage', 'coverage_df_p.tsv'), sep='\t', )

#Add the expression data to it
h_contig_df = h_contig_df.join(rnaseq_df_dict['Pst_104E_v13_h_gene_rpkm_average.tsv'].set_index('name'), on='name').copy()
p_contig_df = p_contig_df.join(rnaseq_df_dict['Pst_104E_v13_p_gene_rpkm_average.tsv'].set_index('name'), on='name').copy()

#Save out the dataframe
h_contig_df.to_csv(os.path.join(DIRS['FIGURES'], 'expression', 'rnaseq_methyl_df_h.tsv'), header=True, index=None, sep='\t')
p_contig_df.to_csv(os.path.join(DIRS['FIGURES'], 'expression', 'rnaseq_methyl_df_p.tsv'), header=True, index=None, sep='\t')

#### <span style='color:#a347ff'> 7.A.2 Run Spearman stats on these dataframes.<span/>

# make a list of observed columns for Spearman
column_list = [ '5mC_gene_body', '6mA_gene_body', '5mC_upstream', '6mA_upstream', '5mC_downstream', '6mA_downstream', '6mA_tss', '6mA_tss_pb', '5mC_both', '6mA_both']

#print the headings list
column_list

#make a list of randomised columns to see correlation to methylation (put in supplementary)
rand_col = []
for col in column_list:
    rand = col + '_rand'
    rand_col.append(rand)
rand_col

def spearman_r(df, col_list):
    """This function returns a dictionary of spearman's R statistic and p-value for a test of observed and expected sites."""
    spearmanr_dict = {}
    for col in col_list:
        obs = df[col]
        exp = df['UG']
        stat, p = spearmanr(obs, exp)
        spearmanr_dict[col] = stat, p
    return spearmanr_dict

h_spearman = spearman_r(h_contig_df, column_list)
p_spearman = spearman_r(p_contig_df, column_list)

h_spearman

p_spearman

# test out spearman on the randomised methylation
rand_h_spearman = spearman_r(h_contig_df, rand_col)
rand_p_spearman = spearman_r(p_contig_df, rand_col)

# these have correlations of 0 :///

rand_h_spearman

rand_p_spearman

def save_spearman_dict(wdict, genome):
    df = pd.DataFrame.from_dict(wdict, orient='index')
    df.rename(columns={0: "Spearman's rho", 1: 'p-value'}, inplace=True)
    out_fn = os.path.join(DIRS['FIGURES'], 'expression', ('spearman_r_table_%s.tsv' % genome))
    df.to_csv(out_fn, header=True, index=True, sep = '\t')

save_spearman_dict(h_spearman, 'h')
save_spearman_dict(p_spearman, 'p')

#save out the null one as well
save_spearman_dict(rand_h_spearman, 'rand_h')
save_spearman_dict(rand_p_spearman, 'rand_p')

### <span style='color:#15c66e'> 7.B Methylated transposons with genes. <span/>

#### <span style='color:#a347ff'> 7.B.1 Making dataframe with methylation + expression data + gene TE data.<span/>

# add columns for genes close to highly methylated transposons and unmethylated transposons

# make lists of the genes close to transposons with a high level of each type of methylation

# make a list of file names
gt_fn_dict = file_name_dict([fn for fn in glob.iglob('%s/*bed' % os.path.join(DIRS['GENE_TE']), recursive=True)])

gt_fn_dict

# make a header list for converting the bed files to dataframes
header_list_gt = ['contig', 'start', 'stop', 'name', 'score', 'gene', 'distance']

def gene_te_df_dict(file_dict, header_row, header_list):
    """Outputs a dictionary of dataframes from a dictionary of file names."""
    df_dict = {}
    for key, value in file_dict.items():
        df = pd.read_csv(value, sep='\t', header=header_row, names=header_list_gt)
        df = df.loc[:,['gene', 'distance']].sort_values('gene', ascending=True)
        df_dict[key] = df
    return df_dict

gt_df_dict = gene_te_df_dict(gt_fn_dict, None, header_list_gt)

print(*gt_df_dict, sep='\n')

def list_gene_names(df):
    gene_list = []
    if df['gene'] != '.':
        gene_list.append(df['gene'])

#make a list of genes in each type
m5c_top_h = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_h_ctg.REPET.5mC_hc_tombo_sorted.cutoff.0.80.overlap.500.top.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_top_h = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_h_ctg.REPET.6mA_hc_tombo_sorted.cutoff.0.80.overlap.500.top.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_pb_top_h = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_h_ctg.REPET.6mA_prob_smrtlink_sorted.cutoff.0.80.overlap.500.top.sorted.bed']), sep='\t', header=None)[5].tolist()))
m5c_low_h = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_h_ctg.REPET.5mC_hc_tombo_sorted.cutoff.0.80.overlap.500.low.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_low_h = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_h_ctg.REPET.6mA_hc_tombo_sorted.cutoff.0.80.overlap.500.low.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_pb_low_h = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_h_ctg.REPET.6mA_prob_smrtlink_sorted.cutoff.0.80.overlap.500.low.sorted.bed']), sep='\t', header=None)[5].tolist()))

m5c_top_p = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_p_ctg.REPET.5mC_hc_tombo_sorted.cutoff.0.80.overlap.500.top.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_top_p = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_p_ctg.REPET.6mA_hc_tombo_sorted.cutoff.0.80.overlap.500.top.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_pb_top_p = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_p_ctg.REPET.6mA_prob_smrtlink_sorted.cutoff.0.80.overlap.500.top.sorted.bed']), sep='\t', header=None)[5].tolist()))
m5c_low_p = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_p_ctg.REPET.5mC_hc_tombo_sorted.cutoff.0.80.overlap.500.low.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_low_p = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_p_ctg.REPET.6mA_hc_tombo_sorted.cutoff.0.80.overlap.500.low.sorted.bed']), sep='\t', header=None)[5].tolist()))
m6a_pb_low_p = list(set(pd.read_csv(os.path.join(gt_fn_dict['Pst_104E_v13_p_ctg.REPET.6mA_prob_smrtlink_sorted.cutoff.0.80.overlap.500.low.sorted.bed']), sep='\t', header=None)[5].tolist()))

h_contig_df['Top_5mC_Gene_TE'] = 'Other'
h_contig_df.loc[h_contig_df.name.isin(m5c_top_h), 'Top_5mC_Gene_TE'] = 'Methylated\nTE adjacent'

h_contig_df['Top_6mA_Gene_TE'] = 'Other'
h_contig_df.loc[h_contig_df.name.isin(m6a_top_h), 'Top_6mA_Gene_TE'] = 'Methylated\nTE adjacent'

h_contig_df['Top_6mA_pb_Gene_TE'] = 'Other'
h_contig_df.loc[h_contig_df.name.isin(m6a_pb_top_h), 'Top_6mA_pb_Gene_TE'] = 'Methylated\nTE adjacent'

h_contig_df['Low_5mC_Gene_TE'] = 'Other'
h_contig_df.loc[h_contig_df.name.isin(m5c_low_h), 'Low_5mC_Gene_TE'] = 'Methylated\nTE adjacent'

h_contig_df['Low_6mA_Gene_TE'] = 'Other'
h_contig_df.loc[h_contig_df.name.isin(m6a_low_h), 'Low_6mA_Gene_TE'] = 'Methylated\nTE adjacent'

h_contig_df['Low_6mA_pb_Gene_TE'] = 'Other'
h_contig_df.loc[h_contig_df.name.isin(m6a_pb_low_h), 'Low_6mA_pb_Gene_TE'] = 'Methylated\nTE adjacent'

p_contig_df['Top_5mC_Gene_TE'] = 'Other'
p_contig_df.loc[p_contig_df.name.isin(m5c_top_p), 'Top_5mC_Gene_TE'] = 'Methylated\nTE adjacent'

p_contig_df['Top_6mA_Gene_TE'] = 'Other'
p_contig_df.loc[p_contig_df.name.isin(m6a_top_p), 'Top_6mA_Gene_TE'] = 'Methylated\nTE adjacent'

p_contig_df['Top_6mA_pb_Gene_TE'] = 'Other'
p_contig_df.loc[p_contig_df.name.isin(m6a_pb_top_p), 'Top_6mA_pb_Gene_TE'] = 'Methylated\nTE adjacent'

p_contig_df['Low_5mC_Gene_TE'] = 'Others'
p_contig_df.loc[p_contig_df.name.isin(m5c_low_p), 'Low_5mC_Gene_TE'] = 'Methylated\nTE adjacent'

p_contig_df['Low_6mA_Gene_TE'] = 'Other'
p_contig_df.loc[p_contig_df.name.isin(m6a_low_p), 'Low_6mA_Gene_TE'] = 'Methylated\nTE adjacent'

p_contig_df['Low_6mA_pb_Gene_TE'] = 'Other'
p_contig_df.loc[p_contig_df.name.isin(m6a_pb_low_p), 'Low_6mA_pb_Gene_TE'] = 'Methylated\nTE adjacent'

#Save out the dataframe
h_contig_df.to_csv(os.path.join(DIRS['FIGURES'], 'expression', 'rnaseq_methyl_df_h.tsv'), header=True, index=None, sep='\t')
p_contig_df.to_csv(os.path.join(DIRS['FIGURES'], 'expression', 'rnaseq_methyl_df_p.tsv'), header=True, index=None, sep='\t')
