==> fisher/methylation/9_BLAST_result_expression_through_life_cycle.ipynb <==
# <span style="color:#ff1414"> BEDtools analysis. </span>

This is a script to answer research questions outlined elsewhere. In summary, this script:

1. compares methylation results between different methylation-callers, and between different methylation sequencing methods.

2. compares methylation between genes and non-gene regions

3. compares methylation between transposons and non-repetitive regions

4. compares transposons and genes


Note:
- PB/pb = PacBio
- ONT/ont = Oxford Nanopore Technology
- NP = Nanopolish

# load modules
import os
import glob
import pprint
import pandas as pd
from scipy import stats
from statsmodels.stats.multicomp import pairwise_tukeyhsd
from statsmodels.stats.multicomp import MultiComparison
import numpy as np

#First we need to define the base dirs
DIRS = {}
DIRS['BASE2'] = '/home/anjuni/analysis'
DIRS['RNA'] = os.path.join(DIRS['BASE2'], 'rna_counts')
DIRS['MEAN_STD'] = os.path.join(DIRS['RNA'], 'average_and_stdev')
DIRS['TRIALS'] = os.path.join(DIRS['RNA'], 'trials_tsv')
DIRS['FIGURES'] = os.path.join(DIRS['BASE2'], 'figures')
DIRS['BLAST_FIG'] = os.path.join(DIRS['FIGURES'], 'blast_expression')

#Quick chech if directories exist
for value in DIRS.values():
    if not os.path.exists(value):
        print('%s does not exist' % value)

## <span style='color:#ffa347'> 10. Expression of methylation machinery throughout Pst life cycle. <span/>

##### candidate protein details copied from BLAST.ipnyb


#### DNMT1/MASC2

<span style='color:darkred'> query: XP_001833175.2, len = 1253 <span/>

<span style='color:purple'> h_subject: Pst104E_20230, len = 1248 <span/>

Score = 206 bits (523),  Expect = 5e-54, Method: Compositional matrix adjust.
Identities = 241/925 (26%), Positives = 372/925 (40%), Gaps = 178/925 (19%)


<span style='color:darkblue'> p_subject: Pst104E_04293, len = 1248 <span/>

Score = 206 bits (523),  Expect = 5e-54, Method: Compositional matrix adjust.
Identities = 241/925 (26%), Positives = 372/925 (40%), Gaps = 178/925 (19%)

#### RAD8

<span style='color:darkred'> query: XP_001831325.2, len = 2184 <span/>

<span style='color:purple'> h_subject: Pst104E_28179, len = 2204 <span/>

Score = 1158 bits (2996),  Expect = 0.0, Method: Compositional matrix adjust.
Identities = 643/1386 (46%), Positives = 837/1386 (60%), Gaps = 76/1386 (5%)



<span style='color:darkblue'> p_subject: Pst104E_12497, len = 2204 <span/>

Score = 1159 bits (2997),  Expect = 0.0, Method: Compositional matrix adjust.
Identities = 643/1386 (46%), Positives = 838/1386 (60%), Gaps = 76/1386 (5%)

DNMT1/MASC2
- h_subject: Pst104E_20230   //   gene_model_hcontig_0009_24.226
- p_subject: Pst104E_04293   //   gene_model_pcontig_009.363
    
RAD8
- h_subject: Pst104E_28179   //   gene_model_hcontig_0052_06.60
- p_subject: Pst104E_12497   //   gene_model_pcontig_052.64


# make dicts of BLAST results
blast_h = { 'DNMT1/MASC2' : 'gene_model_hcontig_0009_24.226', 'RAD8' : 'gene_model_hcontig_0052_06.60' }
blast_p = { 'DNMT1/MASC2' : 'gene_model_pcontig_009.363', 'RAD8' : 'gene_model_pcontig_052.64', 'ALKBH1' : 'gene_model_pcontig_043.133'}

# make dict of all BLAST results
blast_dict = { 'DNMT1/MASC2_h' : 'gene_model_hcontig_0009_24.226', 'RAD8_h' : 'gene_model_hcontig_0052_06.60', \
              'DNMT1/MASC2_p' : 'gene_model_pcontig_009.363', 'RAD8_p' : 'gene_model_pcontig_052.64' }

blast_dict

# make fn dict
rnaseq_fn_dict = {}
for fn in glob.iglob('%s/*_gene_rpkm_average.tsv' % DIRS['MEAN_STD'], recursive=True):
    if len(fn) == 87:
        rnaseq_fn_dict[fn.split('/')[-1]] = fn

rnaseq_fn_dict        

# make rnaseq df dict
rnaseq_df_dict = {}
for fn in glob.iglob('%s/*_gene_rpkm_average.tsv' % DIRS['MEAN_STD'], recursive=True):
    if len(fn) == 87:
        df = pd.read_csv(fn, header=0, sep='\t')
        rnaseq_df_dict[fn.split('/')[-1]] = df

print(*rnaseq_df_dict, sep='\t')

# check df format
rnaseq_df_dict['Pst_104E_v13_h_gene_rpkm_average.tsv'].head()

# write function to pull line for each gene
def get_gene_expression(rnaseq_df, gene_name_dict):
    """This function makes a list of dataframes with expression data for the provided dictionary of genes."""
    df_list = []
    for key, value in gene_name_dict.items():
        df = rnaseq_df.loc[rnaseq_df['gene_ID'] == value]
        df_list.append(df)
    return df_list

# make final df for p and h contig
h_exp_df = pd.concat(get_gene_expression(rnaseq_df_dict['Pst_104E_v13_h_gene_rpkm_average.tsv'], blast_h))
p_exp_df = pd.concat(get_gene_expression(rnaseq_df_dict['Pst_104E_v13_p_gene_rpkm_average.tsv'], blast_p))

# check df format
h_exp_df

# check df format
p_exp_df

# save out the files
def save_tsv(df, file_name):
    out_fn = os.path.join(DIRS['BLAST_FIG'], file_name)
    df.to_csv(out_fn, sep='\t', header=True, index=None)
    
save_tsv(h_exp_df, 'average_exp_h.tsv')
save_tsv(p_exp_df, 'average_exp_p.tsv')

# getting all values, not just averages as before

# make a dict of file paths
trials_fn_dict = {}
for fn in glob.iglob('%s/*_gene_repRpkmMatrix_featureCounts.tsv' % DIRS['TRIALS'], recursive=True):
    if len(fn) == 95:
        trials_fn_dict[fn.split('/')[-1]] = fn

trials_fn_dict

# make trials df dict
trials_df_dict = {}
for fn in glob.iglob('%s/*_gene_repRpkmMatrix_featureCounts.tsv' % DIRS['TRIALS'], recursive=True):
    if len(fn) == 95:
        df = pd.read_csv(fn, header=0, sep='\t')
        trials_df_dict[fn.split('/')[-1]] = df

print(*trials_df_dict)

# make final df for p and h contig
h_trials_df = pd.concat(get_gene_expression(trials_df_dict['Pst_104E_v13_h_gene_repRpkmMatrix_featureCounts.tsv'], blast_h))
p_trials_df = pd.concat(get_gene_expression(trials_df_dict['Pst_104E_v13_p_gene_repRpkmMatrix_featureCounts.tsv'], blast_p))

# check df format
h_trials_df

# check df format
p_trials_df

# save out the dataframes
save_tsv(h_trials_df, 'trials_exp_h.tsv')
save_tsv(p_trials_df, 'trials_exp_p.tsv')

def anova(df):
    f, p = stats.f_oneway(df['UG'], df['GS'], df['IT6'], df['IT9'], df['HE'])
    return f, p

a_df = pd.read_csv(os.path.join(DIRS['FIGURES'], 'blast_expression', 'alkbh1.txt'), header=0, sep='\t')
r_df = pd.read_csv(os.path.join(DIRS['FIGURES'], 'blast_expression', 'rad8.txt'), header=0, sep='\t')
d_df = pd.read_csv(os.path.join(DIRS['FIGURES'], 'blast_expression', 'dnmt1_masc2.txt'), header=0, sep='\t')

print(anova(a_df))
print(anova(r_df))
print(anova(d_df))

a_df

r_df # rad8

d_df # dnmt1/masc2

a_array = a_df.values.flatten()
print(a_array)
a_col = np.array(list(a_df.columns)*3).reshape(3,5).flatten()
print(a_col)

d_array = d_df.values.flatten()
print(d_array)
d_col = np.array(list(d_df.columns)*3).reshape(3,5).flatten()
print(d_col)

r_array = r_df.values.flatten()
print(r_array)
r_col = np.array(list(r_df.columns)*3).reshape(3,5).flatten()
print(r_col)

mc = MultiComparison(r_array.flatten(), r_col.flatten())
result = mc.tukeyhsd()
 
print(result)
print(mc.groupsunique)

# save out results
df = pd.DataFrame(data=result._results_table.data[1:], columns=result._results_table.data[0])
df.to_csv(os.path.join(DIRS['FIGURES'], 'blast_expression', 'rad8_hsd.tsv'), header=True, index=False, sep='\t')

mc = MultiComparison(d_array.flatten(), d_col.flatten())
result = mc.tukeyhsd()
 
print(result)

# save out results
df = pd.DataFrame(data=result._results_table.data[1:], columns=result._results_table.data[0])
df.to_csv(os.path.join(DIRS['FIGURES'], 'blast_expression', 'dnmt1_masc2_hsd.tsv'), header=True, index=False, sep='\t')

mc = MultiComparison(a_array.flatten(), a_col.flatten())
result = mc.tukeyhsd()
 
print(result)

# save out results
df = pd.DataFrame(data=result._results_table.data[1:], columns=result._results_table.data[0])
df.to_csv(os.path.join(DIRS['FIGURES'], 'blast_expression', 'alkbh1_hsd.tsv'), header=True, index=False,  sep='\t')
